Extract rainfall data

Extract rainfall data

ex_rainfall(
  data_path = "\\\\projectdata.eurac.edu/projects/Proslide/PREC_GRIDS_updated/",
  spatial.obj = NULL,
  fun = NULL,
  date = NULL,
  nc_var = "precipitation",
  days_back = 2
)

Arguments

data_path
Path to the gridded rainall data

fun
A function as character that the exactext_extract-function accepts

date
A character that can be coerced directly to a date or an object of class date

nc_var
The variable in the NetCDF you want to extract

days_back
integer vector of days of antecedent rainfall you want to extract

sptial.obj
An object of type sf. Either a point (or MULTIPOINT) or a polygon (or MULTIPOLYGON))

Value

a dataframe with either point extractions or columns of the aggregated values